==> 07-conclusion.Rmd <==
# Conclusion

After our exploratory data analysis, we can conclude that the CO2 concentration consistently increases over these years in general. It can be affected by many factors. Given the data sources we have, time, location, and sea water features can affect the CO2 concentration. It follows a pattern that the CO2 concentration is higher in summer but relatively lower in spring and autumn in both northern and southern hemispheres. So the monthly CO2 concentration in the northern and southern hemispheres have completely different patterns. The northern hemisphere has its peak in May, while the southern hemisphere has its peak in January. On the other hand, even though there are 7 sampling stations that [*Scipps CO2 Program*](https://scrippsco2.ucsd.edu) collecting the CO2 data, this does not have significant effects on the CO2 concentration. But having more samples helps us generate more accurate observations. In addition, the higher the salinity of seawater is, the higher the content of soluble inorganic carbon (DIC) is. Although we suspect that COVID-19 had an effect on CO2 concentration, we find that the effect is so subtle that it can be negligible for the past one and a half years. However, since there is still a slight difference, as COVID-19 becomes a regular event, this difference will increase over time, we might expect CO2 concentrations to get lower in the long run.

On the other hand, there are limitations in this project. The raw data sources in [Scipps CO2 Program](https://scrippsco2.ucsd.edu) have very few categorical variables. So it was challenging for us at first to generate an analysis on different features. There were very limited choices of types of graphs we could plot on. To solve this problem, we manually add a few categories to the data. For example, we add a variable to represent the hemisphere based on the latitude of inspection stations. And we also add a category representing whether the date is in Covid-19 time period. To have more comprehensive perspectives on CO2 emission, more categories and dimensions might be needed. Also, as shown in the map in Chapter1:Introduction, the sampling stations are mainly located in a similar longitude. This might create a limitation to this project as well since we lack the information on CO2 that is collected from other various locations.

In the future, we can still gather more data sources regarding to the CO2 concentration. Besides [Scipps CO2 Program](https://scrippsco2.ucsd.edu), more data can be used for future analysis. The emission of CO2 is a wide-ranging topic. It can be related to many other factors such as chemicals, elevations, population, technology, human activities, etc. Even for the topic of Covid-19, we can continue to monitor the impact of the COVID-19 on CO2 concentrations, as this is just the beginning. And [Scipps CO2 Program](https://scrippsco2.ucsd.edu) is a program that is consistently updating the data. More analysis can be made with future observations.
